# pkg/graphql/models_gen.go
// Code generated by github.com/99designs/gqlgen, DO NOT EDIT.

package graphql

type User struct {
	ID          string `json:"id"`
	Name        string `json:"name"`
	Surname     string `json:"surname"`
	Patronymic  string `json:"patronymic"`
	Age         int    `json:"age"`
	Gender      string `json:"gender"`
	Nationality string `json:"nationality"`
}